% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_CNN_binary.R
\name{deploy_CNN_binary}
\alias{deploy_CNN_binary}
\title{Transfer Learning from Sound Directories}
\usage{
deploy_CNN_binary(
  output_folder,
  output_folder_selections,
  output_folder_wav,
  top_model_path,
  path_to_files,
  detect_pattern = NA,
  architecture,
  clip_duration = 12,
  hop_size = 6,
  downsample_rate = 16000,
  threshold = 0.5,
  save_wav = TRUE,
  positive.class = "Gibbons",
  negative.class = "Noise",
  max_freq_khz = 2
)
}
\arguments{
\item{output_folder}{A character string specifying the path to the output folder where the results will be saved.}

\item{output_folder_selections}{A character string specifying the path to the folder where selection tables will be saved.}

\item{output_folder_wav}{A character string specifying the path to the folder where extracted WAV files will be saved.}

\item{top_model_path}{A character string specifying the path to the pre-trained top model for classification.}

\item{path_to_files}{A character string specifying the path to the directory containing sound files to process.}

\item{clip_duration}{The duration of each sound clip in seconds.}

\item{hop_size}{The hop size for splitting the sound clips.}

\item{downsample_rate}{The downsample rate for audio in Hz, set to 'NA' if no downsampling is required.}

\item{threshold}{The threshold for audio detection.}

\item{save_wav}{A logical value indicating whether to save the extracted sound clips as WAV files.}

\item{max_freq_khz}{The maximum frequency in kHz for spectrogram visualization.}

\item{class_names}{A character vector containing the unique classes for training the model.}

\item{noise_category}{A character string specifying the noise category for exclusion.}
}
\description{
This function processes sound data from a specified directory, performs transfer learning using a pre-trained deep learning model, and saves the results.
}
\details{
This function processes sound data from a directory, extracts sound clips, converts them to images, performs image classification using a pre-trained deep learning model, and saves the results including selection tables and image and audio files.
}
\examples{
{
#' Load data
data("TempBinWav")

dir.create(paste(tempdir(),'/BinaryDir/Wav/'),recursive = T, showWarnings = FALSE)

#' Write to temp directory
writeWave(TempBinWav,filename = paste(tempdir(),'/BinaryDir/Wav/','TempBinWav.wav'))

train_CNN_binary(
  input.data.path = "inst/extdata/binary/",
  test.data = "inst/extdata/binary/test/",
  architecture = "alexnet",  #' Choose 'alexnet', 'vgg16', 'vgg19', 'resnet18', 'resnet50', or 'resnet152'
  unfreeze.param = TRUE,
  batch_size = 6,
  learning_rate = 0.001,
  epoch.iterations = 1,  #' Or any other list of integer epochs
  early.stop = "yes",
  save.model= TRUE,
  output.base.path = paste(tempdir(),'/BinaryDir/',sep=''),
  trainingfolder = "test_binary"
)

TempFileList <- list.files(paste(tempdir(),'/BinaryDir/',sep=''),full.names = T,recursive = T)
ModelPath <- TempFileList[which(str_detect(TempFileList,'.pt'))]


deploy_CNN_binary (
  clip_duration = 12,
  architecture='alexnet',
  output_folder = paste(tempdir(),'/BinaryDir/Results/Images/',sep=''),
  output_folder_selections = paste(tempdir(),'/BinaryDir/Results/Selections/',sep=''),
  output_folder_wav = paste(tempdir(),'/BinaryDir/Results/Wavs/',sep=''),
  detect_pattern=NA,
  top_model_path = ModelPath,
  path_to_files = paste(tempdir(),'/BinaryDir/Wav/'),
  downsample_rate = 'NA',
  threshold = 0.5,
  save_wav = F,
  positive.class = 'Gibbons',
  negative.class = 'Noise',
  max_freq_khz = 2
)
}
}
