% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_trainedmodel_performance.R
\name{evaluate_trainedmodel_performance}
\alias{evaluate_trainedmodel_performance}
\title{Evaluate Model Performance on Image Data}
\usage{
evaluate_trainedmodel_performance(
  trained_models_dir,
  image_data_dir,
  output_dir = "data/",
  positive.class = "Gibbons",
  negative.class = "Noise"
)
}
\arguments{
\item{trained_models_dir}{Path to the directory containing trained models (.pt files).}

\item{image_data_dir}{Path to the directory containing image data for evaluation.}

\item{output_dir}{Path to the directory where the performance scores will be saved.}
}
\value{
Invisible NULL. The performance scores are written to the specified output directory.
}
\description{
Given trained models and a set of images, this function evaluates the performance of the models.
}
