% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_best_performance.R
\name{get_best_performance}
\alias{get_best_performance}
\title{Extract Best Performance Results from Performance Tables}
\usage{
get_best_performance(
  performancetables.dir,
  model.type = "multi",
  class = "hornbill.helmeted",
  Thresh.val = 0.5
)
}
\arguments{
\item{performancetables.dir}{Path to the directory containing the performance tables.}
}
\value{
A list containing best F1 scores, best precision results, best recall results,
and plots visualizing these metrics.
}
\description{
Given the path to a directory of performance tables, this function reads in the tables,
combines them, and extracts the best performance results based on various criteria.
}
