% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TempBinWav}
\alias{TempBinWav}
\title{A short .wav file for automated detection tests.}
\format{
\subsection{\code{TempBinWav}}{

A data frame with 7,240 rows and 60 columns:
\describe{
\item{country}{Malaysia}
\item{year}{2024}
...
}
}
}
\source{
\url{https://doi.org/10.5281/zenodo.10927637}
}
\usage{
TempBinWav
}
\description{
A short .wav file for automated detection tests.
}
\keyword{datasets}
