% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_alexnet.R
\name{train_alexnet}
\alias{train_alexnet}
\title{Train the AlexNet Model}
\usage{
train_alexnet(
  input.data.path,
  test.data.path,
  unfreeze = TRUE,
  epoch.iterations,
  early.stop = "yes",
  output.base.path = "data/",
  trainingfolder.short
)
}
\arguments{
\item{input.data.path}{The path to the input data folder.}

\item{test.data.path}{The path to the test data folder.}

\item{unfreeze}{Logical. If TRUE, unfreeze all the layers of the pretrained AlexNet for retraining.
If FALSE, retain the weights of the pretrained layers. Default is TRUE.}

\item{epoch.iterations}{List of integers. The number of epochs for training the model. Default is 20.}

\item{early.stop}{Character. Determines whether early stopping should be applied or not.
"yes" to apply and "no" to skip. Default is 'yes'.}

\item{output.base.path}{Character. The base path where the output files should be saved.
Default is '/Users/denaclink/Desktop/RStudioProjects/Multi-species-detector/data/'.}

\item{trainingfolder.short}{Character. A shortened descriptor of the training data,
used for naming output files. Default is 'imagesmalaysiaHQ'.}
}
\value{
A list containing two elements:
\itemize{
\item \strong{Output_Path}: The path where the model and metadata are saved.
\item \strong{Metadata}: A dataframe containing metadata about the training session.
}
}
\description{
This function is designed to train the AlexNet model on a given dataset.
The model is saved and other metadata is stored for further usage.
}
\examples{
\dontrun{
  train_alexnet("path_to_input_data", "path_to_test_data")
}

}
\seealso{
\code{\link[torch:nn_module]{torch::nn_module()}} and other torch functions.
}
