% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_CNN_binary.R
\name{train_CNN_binary}
\alias{train_CNN_binary}
\title{Train Binary CNN Models}
\usage{
train_CNN_binary(
  input.data.path,
  test.data,
  architecture,
  unfreeze = TRUE,
  batch_size = 32,
  learning_rate,
  save.model = FALSE,
  epoch.iterations = 1,
  early.stop = "yes",
  output.base.path = "data/",
  trainingfolder,
  list.thresholds = seq(0.1, 1, 0.1),
  positive.class = "Gibbons",
  negative.class = "Noise"
)
}
\arguments{
\item{input.data.path}{Character. The path to the input data folder.}

\item{test.data}{Character. The path to the test data folder.}

\item{architecture}{Character. The CNN architecture to use ('alexnet', 'vgg16', 'vgg19', 'resnet18', 'resnet50', or 'resnet152').}

\item{unfreeze}{Logical. Determines if all layers of the pretrained CNN should be unfrozen for retraining.
Default is TRUE.}

\item{batch_size}{Numeric. Batch size for training model.}

\item{learning_rate}{Numeric. The learning rate for training the model.}

\item{save.model}{Logical. Whether to save model for future use.}

\item{epoch.iterations}{List of integers. The number of epochs for training the model. Default is 1.}

\item{early.stop}{Character. Determines whether early stopping should be applied or not.
"yes" to apply and "no" to skip. Default is 'yes'.}

\item{output.base.path}{Character. The base path where the output files should be saved.
Default is 'data/'.}

\item{trainingfolder}{Character. A shortened descriptor of the training data, used for naming output files.}

\item{list.thresholds}{Numerical list indicating thresholds. Default is seq(0.1,1,.1).}

\item{positive.class}{Character. The name of the positive class label. Default is 'Gibbons'.}

\item{negative.class}{Character. The name of the negative class label. Default is 'Noise'.}
}
\value{
A list containing two elements:
\itemize{
\item \strong{Output_Path}: The path where the model and metadata are saved.
\item \strong{Metadata}: A dataframe containing metadata about the training session.
}
}
\description{
This function is designed to train CNN models (alexnet, VGG16, VGG19, ResNet18, ResNet50, or ResNet152) on a given dataset.
The model is saved, and other metadata is stored for further usage.
}
\examples{
\dontrun{
  train_CNN_binary(
    input.data.path = "path_to_input_data",
    test.data = "path_to_test_data",
    architecture = "alexnet",  # Choose 'alexnet', 'vgg16', 'vgg19', 'resnet18', 'resnet50', or 'resnet152'
    unfreeze = TRUE,
    learning_rate = 0.001,
    epoch.iterations = 1,  # Or any other list of integer epochs
    early.stop = "yes",
    output.base.path = "data/",
    trainingfolder = "example_folder_name"
  )
}
}
\seealso{
\code{\link[torch]{nn_module}} and other torch functions.
}
