% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_ResNet18.R
\name{train_ResNet18}
\alias{train_ResNet18}
\title{Train the ResNet18 Model}
\usage{
train_ResNet18(
  input.data.path,
  test.data,
  unfreeze = TRUE,
  learning_rate,
  epoch.iterations = 1,
  early.stop = "yes",
  output.base.path = "data/",
  trainingfolder,
  positive.class = "Gibbons",
  negative.class = "Noise"
)
}
\arguments{
\item{input.data.path}{Character. The path to the input data folder.}

\item{test.data}{Character. The path to the test data folder.}

\item{unfreeze}{Logical. Determines if all layers of the pretrained ResNet18 should be unfrozen for retraining.
Default is TRUE.}

\item{epoch.iterations}{List of integers. The number of epochs for training the model. Default is 1.}

\item{early.stop}{Character. Determines whether early stopping should be applied or not.
"yes" to apply and "no" to skip. Default is 'yes'.}

\item{output.base.path}{Character. The base path where the output files should be saved.
Default is 'data/'.}

\item{trainingfolder}{Character. A shortened descriptor of the training data, used for naming output files.}

\item{positive.class}{Character. The name of the positive class label. Default is 'Gibbons'.}

\item{negative.class}{Character. The name of the negative class label. Default is 'Noise'.}
}
\value{
A list containing two elements:
\itemize{
\item \strong{Output_Path}: The path where the model and metadata are saved.
\item \strong{Metadata}: A dataframe containing metadata about the training session.
}
}
\description{
This function is designed to train the ResNet18 model on a given dataset.
The model is saved, and other metadata is stored for further usage.
}
\examples{
\dontrun{
  train_ResNet18(
    input.data.path = "path_to_input_data",
    test.data = "path_to_test_data",
    unfreeze = TRUE,
    epoch.iterations = 1,  # Or any other list of integer epochs
    early.stop = "yes",
    output.base.path = "data/",
    trainingfolder = "example_folder_name"
  )
}
}
\seealso{
\code{\link[torch]{nn_module}} and other torch functions.
}
