% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_VGG19_multiClass.R
\name{train_VGG19_multiClass}
\alias{train_VGG19_multiClass}
\title{Train a Multi-Class VGG19 Model}
\usage{
train_VGG19_multiClass(
  input.data.path,
  test.data,
  unfreeze = TRUE,
  epoch.iterations = 1,
  early.stop = "yes",
  output.base.path = "data/",
  trainingfolder,
  learning_rate = 0.01,
  noise.category = "noise"
)
}
\arguments{
\item{input.data.path}{Character. The path to the input training data folder. Sub-folders should correspond to class labels.}

\item{test.data}{Character. The path to the input test data folder. Sub-folders should correspond to class labels.}

\item{unfreeze}{Logical. If TRUE, all layers of the pretrained VGG19 will be unfrozen for retraining. Default is TRUE.}

\item{epoch.iterations}{Integer. The number of epochs to train the model for. Default is 1.}

\item{early.stop}{Character. If "yes", early stopping will be applied during training. Default is "yes".}

\item{output.base.path}{Character. The base path where output files will be saved. Default is 'data/'.}

\item{trainingfolder}{Character. A descriptor for naming the output folder.}

\item{learning_rate}{Numeric. Specify the desired learning rate.}
}
\value{
A list containing:
\itemize{
\item \code{Output_Path}: The path where the model and metrics are saved.
\item \code{Test_Eval}: Evaluation metrics on the test data.
}
}
\description{
This function trains an VGG19 model on a given dataset for multi-class image classification.
The trained model, along with performance metrics and other metadata, are saved to disk.
}
\examples{
\dontrun{
  train_VGG19_multiClass(
    input.data.path = "path/to/training/data",
    test.data = "path/to/test/data",
    unfreeze = TRUE,
    epoch.iterations = 10,
    early.stop = "yes",
    output.base.path = "data/output/",
    trainingfolder = "experiment_1"
  )
}
}
\seealso{
\code{\link[torch]{nn_module}}, \code{\link[torch]{dataloader}}, \code{\link[torch]{nn_bce_with_logits_loss}}
}
