% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_embeddings.R
\name{extract_embeddings}
\alias{extract_embeddings}
\title{Extract Embeddings and Create Scatter Plots}
\usage{
extract_embeddings(test_input, model_path, target_class, unsupervised = "TRUE")
}
\arguments{
\item{test_input}{A character string specifying the path to the directory containing the test images.}

\item{model_path}{A character string specifying the path to the pre-trained PyTorch model file.}

\item{target_class}{A character string specifying the class of interest for cluster analysis.}

\item{unsupervised}{Logical, indicates whether to assign 'target_class' to a cluster and calculate NMI and corresponding confusion matrix}
}
\value{
A list containing the following components:
\describe{
\item{EmbeddingsCombined}{A combined scatter plot of embeddings, showing class and cluster colors.}
\item{NMI}{Normalized Mutual Information (NMI) score between clustering results and ground truth labels.}
\item{ConfusionMatrix}{A confusion matrix showing classification performance metrics.}
}
}
\description{
This function loads a fine-tuned Torch model, extracts embeddings from a set of test images,
performs dimensionality reduction using UMAP, and creates scatter plots to visualize the embeddings.
}
\examples{
{
  #' Set model directory
  trained_models_dir <- system.file("extdata", "trainedresnetmulti/", package = "gibbonNetR")

  #' Specify model path
  ModelPath <- list.files(trained_models_dir, full.names = TRUE)

  # Specify model path
  ImageFile <- system.file("extdata", "multiclass/test/", package = "gibbonNetR")

  # Function to extract and plot embeddings
  result <- extract_embeddings(
    test_input = ImageFile,
    model_path = ModelPath,
    target_class = "female.gibbon",
    unsupervised = "TRUE"
  )

  print(result$EmbeddingsCombined)
}
}
