% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrogram_images.R
\name{spectrogram_images}
\alias{spectrogram_images}
\title{Process and Save Spectrogram Images from Sound Files}
\usage{
spectrogram_images(
  trainingBasePath,
  outputBasePath,
  splits,
  minfreq.khz = 0.4,
  maxfreq.khz = 1.6,
  new.sampleratehz = 16000
)
}
\arguments{
\item{trainingBasePath}{Base directory containing the training folders.}

\item{outputBasePath}{Directory where the processed images will be saved.}

\item{splits}{Numeric vector specifying the split ratios for train, valid, and test sets. Defaults to c(0.8, 0.1, 0.1).}

\item{minfreq.khz}{Minimum frequency in kHz for the spectrogram. Defaults to 0.4.}

\item{maxfreq.khz}{Maximum frequency in kHz for the spectrogram. Defaults to 2.}

\item{new.sampleratehz}{New sample rate in Hz for resampling the audio. Defaults to 16000. Set to 'NA' if no resampling is required.}
}
\value{
Invisible NULL
}
\description{
Process and Save Spectrogram Images from Sound Files
}
\examples{
spectrogram_images(
  trainingBasePath = "data/Clips",
  outputBasePath = "data/TrainingImages",
  splits = c(0.7, 0.2, 0.1)
)

}
