% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TempBinWav}
\alias{TempBinWav}
\title{A short .wav file for automated detection tests.}
\format{
\subsection{\code{TempBinWav}}{

This is a 36 second .wav file that was collected using a SwiftOne autonomous recording unit.
The recording has one gibbon female call in it.
\describe{
\item{country}{Malaysia}
\item{year}{2024}
...
}
}
}
\source{
\url{https://doi.org/10.5281/zenodo.10927637}
}
\usage{
TempBinWav
}
\description{
A short .wav file for automated detection tests.
}
\keyword{datasets}
